⚡ Bolt: [performance improvement] Consolidate and optimize snake_case operations
